scraper/requirements.txt
aiofiles==24.1.0
anyio==3.3.4
bcrypt==4.3.0
blinker==1.6.3
click==8.2.1
exceptiongroup==1.3.0
Flask==3.0.3
greenlet==3.2.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
playwright==1.52.0
priority==2.0.0
psycopg2-binary==2.9.10
pydantic==1.10.22
pyee==13.0.0
PyJWT==2.10.1
Quart==0.19.9
quart-cors==0.8.0
schedule==1.2.2
sniffio==1.3.1
SQLAlchemy==2.0.41
SQLAlchemy-Utils==0.41.2
sqlapagination==0.0.1
tenacity==8.0.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.0.6
wsproto==1.2.0
